refactor: 의도가 드러나게 web view 네이밍 변경
