+++ src/js/controller/runWebLotto.js
@@ -1,24 +1,24 @@
 export function runWebLotto(view, game) {
-    view.showPurchaseForm();
+    view.renderPurchaseForm();
 
     view.onPurchaseSubmit((price) => {
         try {
             const result = game.issueLottos(price);
-            view.showIssuedLottos(result);
-            view.showWinningLottoForm();
+            view.renderIssuedLottos(result);
+            view.renderWinningLottoForm();
 
             view.onWinningLottoSubmit(({ winningNumbers, bonusNumber }) => {
                 try {
                     game.setWinningNumbers(winningNumbers);
                     game.setBonusNumber(bonusNumber);
                     const statistics = game.getStatistics();
-                    view.showStatistics(statistics);
+                    view.renderStatistics(statistics);
                 } catch (error) {
-                    view.showError(error);
+                    view.alertError(error);
                 }
             });
         } catch (error) {
-            view.showError(error);
+            view.alertError(error);
         }
     });
 }
